Added Table Utility, Character Creation & Data Saving
Table utility module allows for deep cloning.

Profile cache is used initially for the load character widget.

If the user has not created any characters, then the load character widget is not displayed and the create character widget is displayed instead.

Error messages ae displayed correctly in the text box, cleared on focus, not sent to the server if the user does not focus on the text box before trying to create the character.